Move qpstest.proto to a common area (test/proto) instead of test/cpp/qps
